Revert "Added GAT model and CellLine Convolution General Model (#6)"
This reverts commit eec736294f652fc823e7d6a3f45f1ceb819da13e.
